requirements.txt
adaptivecardbuilder==0.0.9
Flask==1.1.2
Jinja2==2.10.1
ncclient==0.6.7
paramiko==2.7.1
python_crontab==2.6.0
PyYAML==6.0
requests==2.22.0
ruamel.base==1.0.0
webexteamsbot==0.1.4.2
